refactor(shared): promote ReferenceImage schemas to shared contract layer
Both client and server were carrying parallel TypeScript interfaces
for ReferenceImage that had to stay hand-synced. Hoist the Zod
schemas into shared/schemas/asset.schemas.ts and derive the record
types via z.infer on both sides. Server keeps the local ReferenceImageRecord alias for callers that still want a server-local name.

Eliminates a documented drift class without changing wire shapes.
